refactor to eliminate warnings

+++ app/ti5cp/Main.hs
@@ -15,7 +15,7 @@ main = do
     }
 
 defaultHeapSize :: Int
-defaultHeapSize = 2^20
+defaultHeapSize = 2^(20::Int)
 
 defaultThreshold :: Int
 defaultThreshold = 105
